standard-crud-plus-tests

+++ src/main/java/com/example/gostsNaumen/service/document/DocumentService.java
@@ -1,6 +1,9 @@
 package com.example.gostsNaumen.service.document;
 
 import com.example.gostsNaumen.entity.Document;
+import com.example.gostsNaumen.entity.model.StatusEnum;
+import com.example.gostsNaumen.exception.BusinessException;
+import com.example.gostsNaumen.exception.ErrorCode;
 import com.example.gostsNaumen.repository.DocumentRepository;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
@@ -13,9 +16,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class DocumentService {
     private final DocumentRepository documentRepository;
+    private final LifeCycleService lifeCycleService;
 
-    public DocumentService(DocumentRepository documentRepository) {
+    public DocumentService(DocumentRepository documentRepository, LifeCycleService lifeCycleService) {
         this.documentRepository = documentRepository;
+        this.lifeCycleService = lifeCycleService;
     }
     /**
      * Сохранение сущности ГОСТа в БД
@@ -50,16 +55,55 @@ public class DocumentService {
     }
 
     /**
-     * Удаление госта по Id
+     * Удаление ГОСТа по Id
      *
      * @param id id ГОСТа
      */
     @Transactional
     public void deleteDocumentById(Long id) {
         if (id == null) {
-            throw new EntityNotFoundException("Удаление по пустому ID");
+            throw new IllegalArgumentException("Получен null id");
+        }
+        if (!documentRepository.existsById(id)) {
+            throw new EntityNotFoundException("Документа с таким " + id + " не существует");
         }
         getDocumentById(id);
         documentRepository.deleteById(id);
+    }
+
+    /**
+     * Метод для обновления статуса ГОСТа
+     *
+     * @param documentForUpdate документ, у которого будет изменён статус
+     * @param status            новый статус документа {@link StatusEnum}
+     */
+    @Transactional
+    public Document updateDocumentStatus(Document documentForUpdate, StatusEnum status) {
+        if (documentForUpdate.getStatus().equals(status)) {
+            throw new BusinessException(ErrorCode.STATUS_ALREADY_SET);
+        }
+        if (!lifeCycleService.canUpdateStatus(documentForUpdate.getStatus(), status)) {
+            throw new BusinessException(ErrorCode.INVALID_LIFECYCLE_TRANSITION);
+        }
+        documentForUpdate.setStatus(status);
+        return documentForUpdate;
+    }
+
+    /**
+     * Метод для обновления полей ГОСТа
+     *
+     * @param document документ с уже обновлёнными полями, которые нужно сохранить
+     * @return {@code document} – обновлённый документ
+     */
+    @Transactional
+    public Document updateDocument(Document document) {
+
+        Long id = document.getId();
+
+        if (!documentRepository.existsById(id)) {
+            throw new EntityNotFoundException("Документа с таким id: " + id + ", не существует");
+        }
+
+        return documentRepository.save(document);
     }
 }
